requirements.txt
moviepy
PyYAML
modelscope
onnxruntime
open-clip-torch
opencv-python
rotary-embedding-torch
fairscale
imageio-ffmpeg
xformers
pytorch_lightning